adds a few tests for email verification
