+++ src/test/java/io/supertokens/test/EmailPasswordTest.java
@@ -46,19 +46,6 @@ import static org.junit.Assert.*;
  *  - (later) Test that if there are two transactions running with the same password reset token, only one of them
  *  succeed and the other throws ResetPasswordInvalidTokenException, and that there are no more tokens left for that
  *  user.
- *  - When a new user is signed up, check that email verified is false
- *  - Create an email verification token two times, and check that there are two entries in the db for that user with
- *  the right values
- *  - Call the save email verification token with the same token and check that the right error is thrown
- *  - Call the same email verification token with an unknown userId and check that the right error is thrown
- *  - Verify the email successfully, then create an email verification token and check that the right error is thrown.
- *  - Email verify double lock test. First we lock the token table, then the user table. Does this work?
- *  - Do all password reset token tests with email verification token. For example:
- *    - Give invalid token
- *    - Generate two tokens, verify with one token, the other token should throw an invalid token error
- *    - Use an expired token, it should throw an error
- *    - Test the format of the email verification token
- *    - Create token, change email of user, use the token -> should fail with invalid token
  * */
 
 public class EmailPasswordTest {
@@ -558,32 +545,4 @@ public class EmailPasswordTest {
         process.kill();
         assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
     }
-
-    @Test
-    public void verifyEmail() throws Exception {
-        String[] args = {"../"};
-
-        TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
-        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
-
-        if (StorageLayer.getStorage(process.getProcess()).getType() != STORAGE_TYPE.SQL) {
-            return;
-        }
-
-        User user = EmailPassword.signUp(process.getProcess(), "test@example.com", "password");
-
-        assert (!EmailPassword.isEmailVerified(process.getProcess(), user.id));
-
-        String token = EmailPassword.generateEmailVerificationToken(process.getProcess(), user.id);
-
-        assert (token != null);
-
-        EmailPassword.verifyEmail(process.getProcess(), token);
-
-        assert (EmailPassword.isEmailVerified(process.getProcess(), user.id));
-
-        process.kill();
-        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
-    }
-
 }
